Fixed VersionId type

+++ src/SFA.DAS.EmploymentCheck.Api/Responses/RegisterCheckResponse.cs
@@ -2,7 +2,7 @@
 {
     public class RegisterCheckResponse
     {
-        public int? VersionId { get; set; }
+        public short? VersionId { get; set; }
         public string ErrorType { get; set; }
         public string ErrorMessage { get; set; }
     }
